--- a/indexer/packages/postgres/src/db/migrations/migration_files/20250505153439_replace_subaccounts_address_idx_with_address_subaccountnumber_idx.ts
+++ b/indexer/packages/postgres/src/db/migrations/migration_files/20250505153439_replace_subaccounts_address_idx_with_address_subaccountnumber_idx.ts
@@ -1,0 +1,28 @@
+import * as Knex from 'knex';
+
+export async function up(knex: Knex): Promise<void> {
+  await knex.raw(`
+    DROP INDEX CONCURRENTLY IF EXISTS subaccounts_address_index;
+  `);
+
+  await knex.raw(`
+    CREATE INDEX CONCURRENTLY IF NOT EXISTS subaccounts_address_subaccountnumber_index
+      ON subaccounts(address, "subaccountNumber")
+      INCLUDE (id);
+  `);
+}
+
+export async function down(knex: Knex): Promise<void> {
+  await knex.raw(`
+    DROP INDEX CONCURRENTLY IF EXISTS subaccounts_address_subaccountnumber_index;
+  `);
+
+  await knex.raw(`
+    CREATE INDEX CONCURRENTLY IF NOT EXISTS subaccounts_address_index
+      ON subaccounts (address);
+  `);
+}
+
+export const config = {
+  transaction: false,
+};