### Changelist
generate series takes a long time. save in absolute latency is small but this is invoked for every parent subaccount query

previous query
- without new index
  - address with 125 subaccounts: ~0.7ms
  - address with 40 subaccounts: ~0.67ms
- with new index: ~0.6ms

new query
- without new index:
  - address with 125 subaccounts: ~0.16ms
  - address with 40 subaccounts: ~0.07ms
  - address with 20 subaccounts: ~0.05ms
- with new index: ~0.08ms

### Test Plan
- existing unit tests on correctness of this query
- tested on internal mainnet FE and indexer API response

### Author/Reviewer Checklist
- [ ] If this PR has changes that result in a different app state given the same prior state and transaction list, manually add the `state-breaking` label.
- [ ] If the PR has breaking postgres changes to the indexer add the `indexer-postgres-breaking` label.
- [ ] If this PR isn't state-breaking but has changes that modify behavior in `PrepareProposal` or `ProcessProposal`, manually add the label `proposal-breaking`.
- [ ] If this PR is one of many that implement a specific feature, manually label them all `feature:[feature-name]`.
- [ ] If you wish to for mergify-bot to automatically create a PR to backport your change to a release branch, manually add the label `backport/[branch-name]`.
- [ ] Manually add any of the following labels: `refactor`, `chore`, `bug`.


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

## Summary by CodeRabbit

- **Bug Fixes**
  - Improved database query logic for subaccount retrieval, resulting in more efficient filtering and performance.
- **Chores**
  - Updated database indexing strategy for subaccounts to enhance query speed and reliability.
  - Added descriptive comments to constants for better code clarity and maintainability.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->